Created initial threads and tags views
I began working on an individual views for threads and tags. At the moment it
only directs the user to filler pages, eventually it should show users a
particular thread, threads matching a tag, or a 404 message.

I also changed how the development site redirects the user. It now directs to
the hash_board site when the user enters a url beginning with 'board'. This is
just a quick workaround to deal with the regular expression eating the other
pages.