Improved the thread model and view
The thread model now holds the tile of the thread along with it's number. It
still does not keep track of the associated posts or tags however.

The view now searches for the corresponding thread and 404's if it's not found.
It displays the thread title and id count. Plans have been made to show
associated posts and tags.