Improved the tag model and view
The tag model now uses the tag contents as the primary key.

The url dispatcher has been improved to scan for characters that would actually
show up in the tag. This allows for better ease of searching for tags for the
user.

The view has been changed to show the name of the tag. I still need to
implement the top threads matching that tag.